Rename files in tftypes.
We're no longer talking about DynamicValues, we're now talking about
Values, rename files accordingly.
